[WIP] Boncy Balls
DOUBLES

+++ Julia/F_kernel.jl
@@ -13,49 +13,49 @@ ctx = cl.Context(dev);
 queue = cl.CmdQueue(ctx); 
 
 const F_kernel = "
-		__kernel void Fimp(__global const float *chargevec,			
-				    __global const float *rprod,
-					__global const float *rsquares,
-				 	__global const float *m,				  
-				  	__global const float *I,
+		__kernel void Fimp(__global const double *chargevec,			
+				    __global const double *rprod,
+					__global const double *rsquares,
+				 	__global const double *m,				  
+				  	__global const double *I,
 				  	__global const int *k,
 					__global const int *l,
-			  	 	__global const float *massvec,			
-				 	__global const float *radvec,
-			  	 	__global const float *massrecip,
-			  	 	__global const float *inertiapair,
-				 	__global const float *rad,	
-				 	__global const float *stuff,			  
-				 	__global float3 *r,
-				 	__global float3 *v,
-				 	__global float3 *w)
+			  	 	__global const double *massvec,			
+				 	__global const double *radvec,
+			  	 	__global const double *massrecip,
+			  	 	__global const double *inertiapair,
+				 	__global const double *rad,	
+				 	__global const double *stuff,			  
+				 	__global double3 *r,
+				 	__global double3 *v,
+				 	__global double3 *w)
 					
 				  
 		{ 			
 			int x = get_global_id(0);
 			int a = k[x] - 1;
 			int b = l[x] - 1;
-		    float d = distance(r[a],r[b]);
-		 	float3 Runit = normalize(r[a] - r[b]);
-			float3 wvec = w[a]*rad[a] + w[b]*rad[b];
-			float3 vtemp = v[a] - v[b]; 
-			float p = dot(vtemp,Runit);
+		    double d = distance(r[a],r[b]);
+		 	double3 Runit = normalize(r[a] - r[b]);
+			double3 wvec = w[a]*rad[a] + w[b]*rad[b];
+			double3 vtemp = v[a] - v[b]; 
+			double p = dot(vtemp,Runit);
 		
-			float collisionflag = 0;
+			double collisionflag = 0;
 			if (d < radvec[x]+0.0001) collisionflag = 1; //0.5*(1 + (radvec[x] - d)/(fabs(radvec[x] - d)));	
-			float c = d + radvec[x];
-			float vw6 = (1/6)*stuff[2]*(pow(c,3)-2*rsquares[x]*c)*(1/pow((c*c - rsquares[x] - rprod[x]),2)-1/pow((c*c - rsquares[x] + rprod[x]),2));
-			float del = rad[a] - rad[b];
-			float fi = (del*(9+14*rad[a])-d*d+2*rad[a]*(13*rad[a]-2))/pow((2*rad[a]+d),7)+(-del*(9+14*rad[b])-d*d+2*rad[b]*(13*rad[b]-2))/pow((2*rad[b]+d),7);
-			float di = (-14*rad[a]*(13*rad[a]-2+7*del)+d*d-63*del)/pow((2*rad[a]+d),8) - 2*d/pow((2*rad[a]+d),7)+(-14*rad[b]*(13*rad[b]-2-7*del)+d*d+63*del)/pow((2*rad[b]+d),8) - 2*d/pow((2*rad[b]+d),7); 
-			float vw121 = stuff[3]*(1/(6*c))*((c*c+8*c*radvec[x]+7*rsquares[x]+28*rprod[x])/pow((c+radvec[x]),8)+(21*rprod[x]+7*radvec[x]+d*d)/pow(d,8)-0.2*di);
-			float vw122 = stuff[3]*(1/(30*c*c))*((c*c+7*c*radvec[x]+6*rsquares[x]+21*rprod[x])/pow((c+radvec[x]),7)+(15*rprod[x]+5*radvec[x]+d*d)/pow(d,7)+fi);
-			float F = ((massvec[x] - chargevec[x])/(d*d) + vw6 - vw121 - vw122)*(1 - collisionflag);				
-			float dF = ((-2*massvec[x] + 2*chargevec[x])/(d*d*d))*(1 - collisionflag)*p;		
+			double c = d + radvec[x];
+			double vw6 = (1/6)*stuff[2]*(pow(c,3)-2*rsquares[x]*c)*(1/pow((c*c - rsquares[x] - rprod[x]),2)-1/pow((c*c - rsquares[x] + rprod[x]),2));
+			double del = rad[a] - rad[b];
+			double fi = (del*(9+14*rad[a])-d*d+2*rad[a]*(13*rad[a]-2))/pow((2*rad[a]+d),7)+(-del*(9+14*rad[b])-d*d+2*rad[b]*(13*rad[b]-2))/pow((2*rad[b]+d),7);
+			double di = (-14*rad[a]*(13*rad[a]-2+7*del)+d*d-63*del)/pow((2*rad[a]+d),8) - 2*d/pow((2*rad[a]+d),7)+(-14*rad[b]*(13*rad[b]-2-7*del)+d*d+63*del)/pow((2*rad[b]+d),8) - 2*d/pow((2*rad[b]+d),7); 
+			double vw121 = stuff[3]*(1/(6*c))*((c*c+8*c*radvec[x]+7*rsquares[x]+28*rprod[x])/pow((c+radvec[x]),8)+(21*rprod[x]+7*radvec[x]+d*d)/pow(d,8)-0.2*di);
+			double vw122 = stuff[3]*(1/(30*c*c))*((c*c+7*c*radvec[x]+6*rsquares[x]+21*rprod[x])/pow((c+radvec[x]),7)+(15*rprod[x]+5*radvec[x]+d*d)/pow(d,7)+fi);
+			double F = ((massvec[x] - chargevec[x])/(d*d) + vw6 - vw121 - vw122)*(1 - collisionflag);				
+			double dF = ((-2*massvec[x] + 2*chargevec[x])/(d*d*d))*(1 - collisionflag)*p;		
 			
-			float3 v_rel = (vtemp - p*Runit) - cross(wvec,Runit);
-			float jf = -collisionflag/(inertiapair[x] + massrecip[x]);
-			float j =  -(1 + stuff[1])*collisionflag*p/massrecip[x];			
+			double3 v_rel = (vtemp - p*Runit) - cross(wvec,Runit);
+			double jf = -collisionflag/(inertiapair[x] + massrecip[x]);
+			double j =  -(1 + stuff[1])*collisionflag*p/massrecip[x];			
 		
 			v[b] += F*Runit*stuff[0]/m[b] + dF*Runit*stuff[0]*stuff[0]/m[b] - j*Runit/m[b] - jf*v_rel/m[b];
 			v[a] += -F*Runit*stuff[0]/m[a] - dF*Runit*stuff[0]*stuff[0]/m[a] + j*Runit/m[a] + jf*v_rel/m[a];
@@ -65,9 +65,9 @@ const F_kernel = "
 "		
 
 const r_kernel = " 
-		__kernel void rstep(__global const float *stuff,
-					__global float3 *v,
-					__global float3 *r)
+		__kernel void rstep(__global const double *stuff,
+					__global double3 *v,
+					__global double3 *r)
 		{
 			int x = get_global_id(0);
 			r[x] += (v[x]*stuff[0]);
