+++ Julia/bouncy_balls_cl.jl
@@ -11,21 +11,21 @@ const r_m = 0.3;
 
 fileread("setup.vec");
 
-global const stuff = float32([0,0,0,0]);
+global const stuff = float64([0,0,0,0]);
 
 try
-	global const warp = float32(settings[6]);	
-	stuff[4] = float32(settings[5]);	
+	global const warp = float64(settings[6]);	
+	stuff[4] = float64(settings[5]);	
 	global const max_step = int32(settings[1]);
-	stuff[1] = float32(settings[2]);
-	stuff[2] = float32(settings[3]);
-	stuff[3] = float32(settings[4]);	
+	stuff[1] = float64(settings[2]);
+	stuff[2] = float64(settings[3]);
+	stuff[3] = float64(settings[4]);	
 catch
 	print("Settings not present or invalid; using defaults\n")
 	global const max_step = int32(5e4);
-	stuff[1] = float32(0.05);
-	stuff[2] = float32(1);
-	global const warp = float32(1);
+	stuff[1] = float64(0.05);
+	stuff[2] = float64(1);
+	global const warp = float64(1);
 end
 
 global t_step = int32(0);
@@ -34,16 +34,16 @@ const n_el = int32(1/2*n*(n-1));
 global n_frames = int32(floor(max_step*stuff[1]*30/warp));
 global framecount = int32(0);
 global tempcount = int32(0);
-global r_tracker = float32(zeros(4,n,n_frames));
+global r_tracker = float64(zeros(4,n,n_frames));
 
-global I = float32(zeros(size(r,2),1));
-global massvec = float32(zeros(n_el,1));
-global chargevec = float32(zeros(n_el,1));
-global radvec = float32(zeros(n_el,1));
-global rsquares = float32(zeros(n_el,1));
-global rprod = float32(zeros(n_el,1));
-global massrecip = float32(zeros(n_el,1));
-global inertiapair = float32(zeros(n_el,1));
+global I = float64(zeros(size(r,2),1));
+global massvec = float64(zeros(n_el,1));
+global chargevec = float64(zeros(n_el,1));
+global radvec = float64(zeros(n_el,1));
+global rsquares = float64(zeros(n_el,1));
+global rprod = float64(zeros(n_el,1));
+global massrecip = float64(zeros(n_el,1));
+global inertiapair = float64(zeros(n_el,1));
 global l1 = int32(zeros(n_el,1));
 global l2 = int32(zeros(n_el,1));
 
@@ -54,11 +54,11 @@ end
 for x = 2:n
 	for y = 1:(x-1)	
 	local i = int32(0.5*(x-1)*(x-2) + y);	
-		massvec[i] = float32(m[x]*m[y]*G);
-		chargevec[i] = float32(q[x]*q[y]*k);
-		radvec[i] = float32(rad[x] + rad[y]);
-		rsquares[i] = float32(rad[x]^2 + rad[y]^2);
-		rprod[i] = float32(2*rad[x]*rad[y]);		
+		massvec[i] = float64(m[x]*m[y]*G);
+		chargevec[i] = float64(q[x]*q[y]*k);
+		radvec[i] = float64(rad[x] + rad[y]);
+		rsquares[i] = float64(rad[x]^2 + rad[y]^2);
+		rprod[i] = float64(2*rad[x]*rad[y]);		
 		massrecip[i] = (1/m[x] + 1/m[y]);
 		inertiapair[i] = rad[y]^2/I[y] + rad[x]^2/I[x];
 		l1[i] = x;
@@ -68,21 +68,21 @@ end
 
 l1buff = cl.Buffer(Int32, ctx, (:r, :copy), hostbuf=l1);
 l2buff = cl.Buffer(Int32, ctx, (:r, :copy), hostbuf=l2);
-cbuff = cl.Buffer(Float32, ctx, (:r, :copy), hostbuf=chargevec);
-m1buff = cl.Buffer(Float32, ctx, (:r, :copy), hostbuf=float32(m));
-m2buff = cl.Buffer(Float32, ctx, (:r, :copy), hostbuf=massvec);
-m3buff = cl.Buffer(Float32, ctx, (:r, :copy), hostbuf=massrecip);
-r1buff = cl.Buffer(Float32, ctx, (:r, :copy), hostbuf=float32(rad));
-r2buff = cl.Buffer(Float32, ctx, (:r, :copy), hostbuf=radvec);
-r3buff = cl.Buffer(Float32, ctx, (:r, :copy), hostbuf=rprod);
-r4buff = cl.Buffer(Float32, ctx, (:r, :copy), hostbuf=rsquares);
-I1buff = cl.Buffer(Float32, ctx, (:r, :copy), hostbuf=float32(I));
-I2buff = cl.Buffer(Float32, ctx, (:r, :copy), hostbuf=inertiapair);
-tbuff = cl.Buffer(Float32, ctx, (:r, :copy), hostbuf=stuff);
+cbuff = cl.Buffer(Float64, ctx, (:r, :copy), hostbuf=chargevec);
+m1buff = cl.Buffer(Float64, ctx, (:r, :copy), hostbuf=float64(m));
+m2buff = cl.Buffer(Float64, ctx, (:r, :copy), hostbuf=massvec);
+m3buff = cl.Buffer(Float64, ctx, (:r, :copy), hostbuf=massrecip);
+r1buff = cl.Buffer(Float64, ctx, (:r, :copy), hostbuf=float64(rad));
+r2buff = cl.Buffer(Float64, ctx, (:r, :copy), hostbuf=radvec);
+r3buff = cl.Buffer(Float64, ctx, (:r, :copy), hostbuf=rprod);
+r4buff = cl.Buffer(Float64, ctx, (:r, :copy), hostbuf=rsquares);
+I1buff = cl.Buffer(Float64, ctx, (:r, :copy), hostbuf=float64(I));
+I2buff = cl.Buffer(Float64, ctx, (:r, :copy), hostbuf=inertiapair);
+tbuff = cl.Buffer(Float64, ctx, (:r, :copy), hostbuf=stuff);
 
-rpbuff = cl.Buffer(Float32, ctx, (:rw, :copy), hostbuf=float32(r_pad));
-vpbuff = cl.Buffer(Float32, ctx, (:rw, :copy), hostbuf=float32(v_pad));
-wpbuff = cl.Buffer(Float32, ctx, (:rw, :copy), hostbuf=float32(w_pad));
+rpbuff = cl.Buffer(Float64, ctx, (:rw, :copy), hostbuf=float64(r_pad));
+vpbuff = cl.Buffer(Float64, ctx, (:rw, :copy), hostbuf=float64(v_pad));
+wpbuff = cl.Buffer(Float64, ctx, (:rw, :copy), hostbuf=float64(w_pad));
 
 p = Progress(max_step,1)
 for t_step = 1:max_step
